split amp/bento base-carousel tests
